Move to normal switch call.
